Fixed warnings for iOS 7/8

+++ StyledPageControlDemo/PageControlDemo/StyledPageControl.h
@@ -49,7 +49,7 @@ typedef enum
 @interface StyledPageControl : UIControl
 @property (nonatomic) UIColor *coreNormalColor, *coreSelectedColor;
 @property (nonatomic) UIColor *strokeNormalColor, *strokeSelectedColor;
-@property (nonatomic, assign) int currentPage, numberOfPages;
+@property (nonatomic, assign) NSInteger currentPage, numberOfPages;
 @property (nonatomic, assign) BOOL hidesForSinglePage;
 @property (nonatomic, assign) PageControlStyle pageControlStyle;
 @property (nonatomic, assign) int strokeWidth, diameter, gapWidth;
